# contracts/openapi/admin/v1/openapi.yaml
openapi: 3.1.0
info:
  title: PadlHub Admin API
  version: 1.2.0
  description: Canonical tenant-scoped CUP boundary for audited operator commands.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.padlhub.ru
security:
  - PadlHubAdminJwt: []
...
paths:
  /admin/api/v1/{tenantKey}/notifications/capabilities:
    get:
      tags: [Notifications]
      operationId: getAdminNotificationCapabilities
      summary: Return the effective notification channels available to CUP
      parameters:
        - $ref: '#/components/parameters/TenantKey'
      responses:
        '200':
          description: Effective channel availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNotificationCapabilities'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
  /admin/api/v1/{tenantKey}/notifications/recipients/resolve:
    post:
      tags: [Notifications]
      operationId: resolveAdminNotificationRecipients
      summary: Resolve active PadlHub recipients by phone number without persisting the input
      parameters:
        - $ref: '#/components/parameters/TenantKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneRecipientRequest'
      responses:
        '200':
          description: Resolved recipients and masked unresolved inputs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNotificationRecipientResolution'
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
  /admin/api/v1/{tenantKey}/notifications/campaigns:
    post:
      tags: [Notifications]
      operationId: createAdminNotificationCampaign
      summary: Create an audited idempotent notification campaign
      parameters:
        - $ref: '#/components/parameters/TenantKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminNotificationCampaign'
      responses:
        '202':
          description: Campaign accepted and projected to inbox and delivery queues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNotificationCampaignAccepted'
        '400':
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '409':
          $ref: '#/components/responses/ApiError'
        '422':
          $ref: '#/components/responses/ApiError'
...
components:
  securitySchemes:
    PadlHubAdminJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: PadlHub-issued staff JWT with the dedicated phub-admin audience.
  parameters:
    TenantKey:
      name: tenantKey
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{1,62}$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  responses:
    ApiError:
      description: Stable PadlHub API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    AdminNotificationChannel:
      type: string
      enum: [IN_APP, WEB_PUSH, IOS_PUSH, ANDROID_PUSH]
    PhoneRecipientRequest:
      type: object
      additionalProperties: false
      required: [phones]
      properties:
        phones:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            minLength: 5
            maxLength: 32
    AdminNotificationChannelCapability:
      type: object
      additionalProperties: false
      required: [channel, enabled]
      properties:
        channel:
          $ref: '#/components/schemas/AdminNotificationChannel'
        enabled:
          type: boolean
        reason:
          type: string
        tenantEnabled:
          type: boolean
    AdminNotificationCapabilities:
      type: object
      additionalProperties: false
      required: [channels]
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/AdminNotificationChannelCapability'
    AdminNotificationRecipient:
      type: object
      additionalProperties: false
      required: [userId, displayName, phoneMasked, availableChannels]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        phoneMasked:
          type: string
        availableChannels:
          type: array
          items:
            $ref: '#/components/schemas/AdminNotificationChannel'
    AdminNotificationRecipientResolution:
      type: object
      additionalProperties: false
      required: [matched, unresolvedPhones]
      properties:
        matched:
          type: array
          items:
            $ref: '#/components/schemas/AdminNotificationRecipient'
        unresolvedPhones:
          type: array
          items:
            type: string
    CreateAdminNotificationCampaign:
      allOf:
        - $ref: '#/components/schemas/PhoneRecipientRequest'
        - type: object
          additionalProperties: false
          required: [title, body, channels]
          properties:
            title:
              type: string
              minLength: 1
              maxLength: 300
            body:
              type: string
              minLength: 1
              maxLength: 8000
            deepLink:
              type: string
              minLength: 1
              maxLength: 2000
              pattern: '^/(?!/).*'
            channels:
              type: array
              minItems: 1
              maxItems: 4
              uniqueItems: true
              items:
                $ref: '#/components/schemas/AdminNotificationChannel'
    AdminNotificationCampaignAccepted:
      type: object
      additionalProperties: false
      required:
        - outcome
        - campaignId
        - matchedCount
        - unresolvedCount
        - inAppCreatedCount
        - pushQueuedCount
        - suppressedCount
        - replayed
      properties:
        outcome:
          type: string
          const: accepted
        campaignId:
          type: string
          format: uuid
        matchedCount:
          type: integer
          minimum: 0
        unresolvedCount:
          type: integer
          minimum: 0
        inAppCreatedCount:
          type: integer
          minimum: 0
        pushQueuedCount:
          type: integer
          minimum: 0
        suppressedCount:
          type: integer
          minimum: 0
        replayed:
          type: boolean
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, correlationId]
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string